pydantic-logfire-try/pyproject.toml
[project]
name = "pydantic-logfire-try"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "logfire[fastapi]>=3.18.0",
    "pydantic-ai-slim[openai]>=0.2.15",
]
